Moved error messages to constant enum
