+++ src/commands/music/save.ts
@@ -3,7 +3,8 @@ import { localizedString, useLocalizedString } from '../../helpers/localization/
 import { PlayerCommand } from '../../models/discord.js';
 
 import getLocalizations from '../../helpers/localization/getLocalizations.js';
-import { logger, DefaultLoggerMessage } from '../../helpers/logger/logger.js';
+import { logger } from '../../helpers/logger/logger.js';
+import { DefaultLoggerMessage } from '../../constants/logger.js';
 import { useDefaultPlayer } from '../../helpers/discord/player.js';
 
 export const Save: PlayerCommand = {
